move expand_fn_transform
