+++ pennylane/transforms/core/transform.py
@@ -15,8 +15,7 @@
 This module contains the transform function/decorator to make your custom transforms compatible with tapes, quantum
 functions and QNodes.
 """
-from functools import wraps
-from typing import get_type_hints, Callable
+from typing import get_type_hints
 from .transform_dispatcher import TransformDispatcher, TransformError
 
 
@@ -188,32 +187,3 @@ def transform(
         is_informative=is_informative,
         final_transform=final_transform,
     )
-
-
-def null_postprocessing(results):
-    """A postprocessing function with null behavior."""
-    return results[0]
-
-
-def expand_fn_transform(expand_fn: Callable) -> "TransformDispatcher":
-    """Construct a transform from a tape-to-tape function.
-
-    Args:
-        expand_fn (Callable): a function from a single tape to a single tape
-
-    Returns:
-
-        .TransformDispatcher: Returns a transform dispatcher object that that can transform any
-        circuit-like object in PennyLane.
-
-    >>> device = qml.device('default.qubit.legacy', wires=2)
-    >>> my_transform = qml.transforms.core.expand_fn_transform(device.expand_fn)
-    >>> my_transform
-    <transform: expand_fn>
-    """
-
-    @wraps(expand_fn)
-    def wrapped_expand_fn(tape, *args, **kwargs):
-        return (expand_fn(tape, *args, **kwargs),), null_postprocessing
-
-    return transform(wrapped_expand_fn)
