+++ tests/workflow/test_construct_batch.py
@@ -24,6 +24,33 @@ import pennylane as qml
 from pennylane.transforms.core.transform_dispatcher import TransformContainer
 from pennylane.transforms.core.transform_program import TransformProgram
 from pennylane.workflow import get_transform_program, construct_batch
+from pennylane.workflow.construct_batch import expand_fn_transform
+
+
+def test_expand_fn_transform():
+    """Tests the expand_fn_transform."""
+
+    def my_expand_fn(tape, op1, op2=qml.S(0), op3=qml.S(0)):
+        """my docstring."""
+        return qml.tape.QuantumScript(
+            tape.operations + [op1, op2, op3], tape.measurements, tape.shots
+        )
+
+    t = expand_fn_transform(my_expand_fn)
+
+    assert isinstance(t, qml.transforms.core.TransformDispatcher)
+    tape = qml.tape.QuantumScript([qml.S(0)], [qml.expval(qml.PauliZ(0))], shots=50)
+
+    batch, fn = t(tape, qml.PauliX(0), op3=qml.T(0))
+    assert len(batch) == 1
+    expected = qml.tape.QuantumScript(
+        [qml.S(0), qml.PauliX(0), qml.S(0), qml.T(0)], [qml.expval(qml.PauliZ(0))], shots=50
+    )
+    assert qml.equal(batch[0], expected)
+    assert fn(("a",)) == "a"
+
+    assert repr(t) == "<transform: my_expand_fn>"
+    assert t.__doc__ == "my docstring."
 
 
 class TestTransformProgramGetter:
